Remove the sidebar (#31)
* Remove the sidebar
With the current implementation, the tags are unusable and distact from the other functionality of the app (which still needs performance tuning). This change removes the side bar to allow for focus on other aspects. It includes a drive-by fix to remove the fade animation when showing the thumbnails (also slow at the moment).

* Update the screenshot.