Remove empty macOS app (#49)
I optimistically added a macOS with a view to supporting a native (not Catalyst) macOS build. Since there's still much work to do to improve the performance and add tests for the iOS app (including adding a first-run wizard, etc), I'm deleting the target for the time being. It should, instead, be re-added as a separate project, and the common code should be extracted into a core library.