Rename the legacy Stuff target to Bookmarks (#50)
The project was originally called 'Stuff', meaning the targets and many of the files were thus named. This change addresses that and renames everything back to 'Bookmarks' to match the current app name.

This change also adds the MIT license header to a couple of source files that had been missed, and deletes empty stub methods from test files.