Enable test builds (#67)
This change enables tests in the GitHub actions build job for the BookmarksCore framework (iOS and macOS), and for the iOS app. In order to do this, it was necessary to drop the macOS deployment target for BookmarksCore to 10.15 as GitHub actions doesn't provide a Big Sur (macOS 11) runner at the moment. Since the planned deployment target for the macOS app is Big Sur, there's no point enabling the tests for this.