Introduce basic functionality to the dedicated macOS app (#71)
This change adds a rudimentary grid view of bookmarks to the macOS app. In order to do so, it moves a little more of the core app functionality out of iOS and into BookmarksCore.framework, and copies BookmarkCell and the guts of BookmarksView from the iOS app into the new macOS target. It will surely make sense to share this code between both apps but while the macOS app is so divergent, copy-and-paste seems safer.