Work around duplicate data in the Pinboard API (#81)
Unfortunately it looks like the Pinboard posts/all API is returning duplicate data, meaning that we're seeing some URLs (and therefore hashes) more than once. Since we're using these for identifiers in the SwiftUI list, it's leading to crashes as it has no idea how to handle duplicate elements. This change pre-filters the data when it's returned from the Pinboard API to ensure bad data never leaks into the app.

This also includes a drive-by fix to clean up the Pinboard class (used for managing the Pinboard API) in advance of adding additional methods.