Save the main window frame (#95)
This change works around SwiftUI's inability to autosave the window frame by injecting an NSView into the SwiftUI view heirarchy and using the lifecycle of this view to set the window frame on construction, and store it in the user defaults when the window frame changes. Hopefully SwiftUI gets official support for this in the future.