ci(macOS): Set up automatic macOS build and release (#104)
This change introduces automated build and release for the macOS app. It adds `changes` (https://github.com/jbmorley/changes) as a submodule for computing version numbers based loosely on the conventional commits / semantic versioning strategy and introduces a further dependency on `build-tools` (https://github.com/jbmorley/build-tools) which provides some lightweight conveniences for manipulating the local keychain and certificate store, validating notarized builds, etc.

In order to simply the binary being build as part of this process, it removes Sparkle from the macOS builds for the time being as it wasn't being used and added unnecessary complexity and bloated the binary.