feat: Show the tags in a sidebar (#154)
This change takes advantage of the new SQLite store to fetch all the tags, present them in the sidebar, and support filtering the bookmarks by tag. It introduces some database fixes to ensure that tags are cleaned up when no more bookmarks exist with that tag, adds rudimentary support for renaming tags, and improves thumbnail performance somewhat.