ci: Update BookmarksCore to iOS 14.5 / macOS 11.3 (#187)
Updating the base iOS and macOS versions of BookmarksCore in advance of moving some of the UI code into the core library to allow UI to be shared between the macOS and iOS apps. This change also removes some unnecessary per-target configuration overrides in an effort to simplify the project configuration.