feat: Modify all active filters (including the section) from the search field (#242)
This change fleshes out the `QueryDescription` protocol and implementations to add quality, type erasure, and round-tripping between textual filters and query descriptions. It also introduces a mechanism to determine the primary section for a given query and will dynamically update the selected sidebar section to correspond with that query, allowing users to change the section by simply editing the filter. It does slightly break the sidebar / detail view model and has to remove the `NavigationLink` and instead use a binding to share the sidebar selection between the two children of the `NavigationView`.