fix: Renamed tags did not update until an app relaunch (#258)
Unfortunately the Pinboard `posts/update` endpoint doesn't report any updates following a tag rename as the posts themselves haven't actually been updated 🤦🏻‍♂️. This change works around that behaviour for the time being by simple forcing a full update following a rename. A better longer-term solution might be to simply not require a fetch from Pinboard in the case of a successful client-side and server-side update (we can probably safely assume that the two stores are in sync at this point).