fix: Show an alert if operations fail (#270)
This code introduces a new `ViewModifier` that injects an error handler into the environment and will present an alert when called with an error. It updates all (or at least as many as I could find) calls where errors were being called with this new error handling code. It also updates some of the `Updater` APIs and the context menu commands to handle sets of items in preparation for item selection.