feat: Quickly add tags to bookmarks from the context menu (#274)
This change introduces a SwiftUI `NSTokenField` wrapper (`TokenField`) and uses this in a new view (`AddTagsView`) to let users quickly select tags with suggestions offered from the existing set of tags. Unfortunately `NSTokenField` only provides a synchronous suggestions API, so this change also introduces a trie to `TagsView` to ensure that prefix string matches can be yielded quickly.