feat: Select and manage multiple bookmarks  (#281)
This change introduces the `SelectionTracker` from Interact and uses this to manage the user selection of multiple items in the list. It also tracks the application focus (to allow the selection color to be desaturated when the app is backgrounded), and introduces a couple of toolbar items to delete and tag bookmarks.