feat: Edit selected bookmarks from the main menu (#288)
This change introduces a new `BookmarksSelection` that's designed to manage and operate on the current active selection, and uses this from the new main menu. Right now it's an app-wide object, but hopefully it can be bound to the window and exposed to the main menu using `FocusedBinding` if this is fixed on the Mac in SwiftUI 3.

In order to be able to reuse the menu items in both the context and main menus, this change adds a new enum (`MenuType`) and injected environment variable to allow menu items to be conditionally decorated with their keyboard shortcuts. This menu type is also used to conditionally disable selection-based enabling/disabling of the menu items when shown in the context menu as this can lead to unreliable behaviour in the menu (items being incorrectly disabled, etc).

Unfortunately there still remains one small issue where the tags submenu doesn't always show on first display of the menu which will need to be addressed in a subsequent change. See https://github.com/inseven/bookmarks/issues/292.