fix: Don't allow the search bar to change the section (#405)
We were previously modelling the currently selected section in the search bar and allowing changes to the search bar to update the section. This doesn't really fit the hierarchical structure of most modern macOS and iOS apps and was working against SwiftUI. This simplifies things so that search only searches the current section.