fix: Prefer platform provided titles in the share extension (#722)
iOS and macOS will often propose a title alongside the URL when sharing
an item to the share extension. This change prefers that title if it's
present to reduce the reliance on fetching the titles ourselves (which
can be unreliable if we're only fetching the page contents and not going
through the full page load in a browser).